+++ src/CommandLine/CommandLineOptions/RenameCommandLineOptions.cs
@@ -11,7 +11,6 @@ namespace Orang.CommandLine;
 
 [Verb("rename", HelpText = "Renames files and directories.")]
 [OptionValueProvider(nameof(Content), OptionValueProviderNames.PatternOptionsWithoutPart)]
-[OptionValueProvider(nameof(Display), OptionValueProviderNames.Display_NonContent)]
 [OptionValueProvider(nameof(Highlight), OptionValueProviderNames.RenameHighlightOptions)]
 [OptionValueProvider(nameof(Name), OptionValueProviderNames.PatternOptionsWithoutGroupAndNegative)]
 [CommandGroup("File System", 1)]
@@ -41,13 +40,6 @@ internal sealed class RenameCommandLineOptions : DeleteOrRenameCommandLineOption
         HelpText = "Display which files/directories should be renamed "
             + "but do not actually rename any file/directory.")]
     public bool DryRun { get; set; }
-
-    [HideFromHelp]
-    [Option(
-        longName: OptionNames.Evaluator,
-        HelpText = "[deprecated] Use option -r, --replacement instead.",
-        MetaValue = MetaValues.Evaluator)]
-    public string Evaluator { get; set; } = null!;
 
     [Option(
         longName: OptionNames.Interactive,
@@ -132,17 +124,6 @@ internal sealed class RenameCommandLineOptions : DeleteOrRenameCommandLineOption
         if (!context.TryParseReplacement(Replacement, out string? replacement, out MatchEvaluator? matchEvaluator))
             return false;
 
-        if (matchEvaluator is null
-            && Evaluator is not null)
-        {
-            context.WriteWarning($"Option '{OptionNames.GetHelpText(OptionNames.Evaluator)}' is has been deprecated "
-                + "and will be removed in future version. "
-                + $"Use option '{OptionNames.GetHelpText(OptionNames.Replacement)}' instead.");
-
-            if (!DelegateFactory.TryCreateFromAssembly(Evaluator, typeof(string), typeof(Match), context.Logger, out matchEvaluator))
-                return false;
-        }
-
         if (!context.TryParseReplaceOptions(
             Modify,
             OptionNames.Modify,
@@ -163,26 +144,18 @@ internal sealed class RenameCommandLineOptions : DeleteOrRenameCommandLineOption
             return false;
         }
 
-        if (!context.TryParseDisplay(
-            values: Display,
-            optionName: OptionNames.Display,
-            contentDisplayStyle: out ContentDisplayStyle? contentDisplayStyle,
-            pathDisplayStyle: out PathDisplayStyle? pathDisplayStyle,
-            lineDisplayOptions: out LineDisplayOptions lineDisplayOptions,
-            lineContext: out LineContext lineContext,
-            displayParts: out DisplayParts displayParts,
-            includeCreationTime: out bool includeCreationTime,
-            includeModifiedTime: out bool includeModifiedTime,
-            includeSize: out bool includeSize,
-            indent: out string? indent,
-            separator: out string? separator,
-            noAlign: out bool _,
-            contentDisplayStyleProvider: OptionValueProviders.ContentDisplayStyleProvider,
-            pathDisplayStyleProvider: OptionValueProviders.PathDisplayStyleProvider_Rename))
-        {
-            return false;
-        }
-
+        var displayParts = DisplayParts.None;
+        var lineDisplayOptions = LineDisplayOptions.None;
+        ContentDisplayStyle? contentDisplayStyle = null;
+        PathDisplayStyle? pathDisplayStyle = null;
+        LineContext lineContext = default;
+#if DEBUG
+        string? indent = null;
+        string? separator = null;
+#else
+        const string? indent = null;
+        const string? separator = null;
+#endif
         if (ContentMode is not null)
         {
             if (context.TryParseAsEnum(
@@ -245,15 +218,6 @@ internal sealed class RenameCommandLineOptions : DeleteOrRenameCommandLineOption
         if (NoPath)
             pathDisplayStyle = PathDisplayStyle.Omit;
 #endif
-        if (includeCreationTime)
-            options.FilePropertyOptions = options.FilePropertyOptions.WithIncludeCreationTime(true);
-
-        if (includeModifiedTime)
-            options.FilePropertyOptions = options.FilePropertyOptions.WithIncludeModifiedTime(true);
-
-        if (includeSize)
-            options.FilePropertyOptions = options.FilePropertyOptions.WithIncludeSize(true);
-
         if (pathDisplayStyle == PathDisplayStyle.Relative
             && options.Paths.Length > 1
             && options.SortOptions is not null)