+++ src/CommandLine/CommandUtility.cs
@@ -132,7 +132,7 @@ internal static class CommandUtility
         if (commandAlias is not null)
         {
             logger.WriteWarning($"Command '{args[0]}' has been deprecated "
-                + $"and will be removed in future version. Use command '{commandAlias}' instead.");
+                + $"and will be removed in future versions. Use command '{commandAlias}' instead.");
         }
 
         args[0] = commandName;