+++ src/CommandLine/Commands/FileSystemCommand`1.cs
@@ -157,6 +157,7 @@ internal abstract class FileSystemCommand<TOptions> : AbstractCommand<TOptions> 
 
         var search = new SearchState(fileMatcher, directoryMatcher)
         {
+            GlobFilter = Options.GlobFilter,
             IncludeDirectory = includeDirectory,
             ExcludeDirectory = excludeDirectory,
             LogProgress = (ProgressReporter is not null) ? p => ProgressReporter.Report(p) : null,
@@ -582,7 +583,8 @@ internal abstract class FileSystemCommand<TOptions> : AbstractCommand<TOptions> 
         long size = -1;
 
         if (Options.IncludeSummary
-            && Options.FilePropertyOptions.IncludeSize)
+            && Options.FilePropertyOptions.IncludeSize
+            && !fileMatch.IsSubmatch)
         {
             size = (fileMatch.IsDirectory)
                 ? (context.DirectorySizeMap?[fileMatch.Path] ?? FileSystemUtilities.GetDirectorySize(fileMatch.Path))
@@ -620,7 +622,11 @@ internal abstract class FileSystemCommand<TOptions> : AbstractCommand<TOptions> 
 
             sb.Append(sizeText);
 
-            context.Telemetry.FilesTotalSize += size;
+            if (!Options.IncludeSummary
+                && !fileMatch.IsSubmatch)
+            {
+                context.Telemetry.FilesTotalSize += size;
+            }
         }
 
         if (Options.FilePropertyOptions.IncludeCreationTime)