+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0-beta] - 2023-10-19
+
+### Added
+
+- Add globbing functionality ([#74](https://github.com/josefpihrt/orang/pull/74)).
+  - Add options `--include`, `--exclude`.
+  - Make option `--include-directory` obsolete.
+
+### Changed
+
+- Update logo ([#72](https://github.com/josefpihrt/orang/pull/72)).
+- Open help in browser on `orang help <COMMAND>` ([#76](https://github.com/josefpihrt/orang/pull/76)).
+  - Deprecate options `--online` and `--filter`.
+- Change traversal to depth-first ([#77](https://github.com/josefpihrt/orang/pull/77)).
+
+### Removed
+
+- Remove deprecated options ([#83](https://github.com/josefpihrt/orang/pull/83)).
+  - Remove option `--display`
+  - Remove option `--evaluator`
+
+## [0.5.0] - 2023-09-26
+
 ### Added
 
 - Add command `regex create` ([#45](https://github.com/josefpihrt/orang/pull/45)).